fix and refactor(editor tiles): Reorganizado, creado y reparado el tile_picker y el tile_toolbar
