+++ maps/combined_map.overlay.json
@@ -860,7 +860,7 @@
     "wal",
     "wal",
     "",
-    "",
+    "fl0",
     "fl1",
     "fl1",
     "fl1",
@@ -1227,7 +1227,7 @@
     "wal",
     "wal",
     "fl0",
-    "fl3",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1471,7 +1471,7 @@
     "wal",
     "wal",
     "fl0",
-    "fl1",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1593,7 +1593,7 @@
     "wal",
     "wal",
     "fl0",
-    "fl1",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1715,7 +1715,7 @@
     "",
     "wal",
     "fl0",
-    "fl1",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
